JitIL: Attempt to fix some 64-bit build errors.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@2204 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/Core/Src/PowerPC/Jit64IL/JitAsm.cpp
@@ -254,7 +254,7 @@ void AsmRoutineManager::GenQuantizedStores() {
 
 	const u8* storePairedU8 = AlignCode4();
 	SHR(32, R(EAX), Imm8(6));
-	MOVSS(XMM1, MDisp(EAX, (u32)m_quantizeTableS));
+	MOVSS(XMM1, MDisp(EAX, (u32)(u64)m_quantizeTableS));
 	PUNPCKLDQ(XMM1, R(XMM1));
 	MULPS(XMM0, R(XMM1));
 	CVTPS2DQ(XMM0, R(XMM0));
@@ -271,7 +271,7 @@ void AsmRoutineManager::GenQuantizedStores() {
 
 	const u8* storePairedS8 = AlignCode4();
 	SHR(32, R(EAX), Imm8(6));
-	MOVSS(XMM1, MDisp(EAX, (u32)m_quantizeTableS));
+	MOVSS(XMM1, MDisp(EAX, (u32)(u64)m_quantizeTableS));
 	PUNPCKLDQ(XMM1, R(XMM1));
 	MULPS(XMM0, R(XMM1));
 	CVTPS2DQ(XMM0, R(XMM0));
@@ -288,7 +288,7 @@ void AsmRoutineManager::GenQuantizedStores() {
 
 	const u8* storePairedU16 = AlignCode4();
 	SHR(32, R(EAX), Imm8(6));
-	MOVSS(XMM1, MDisp(EAX, (u32)m_quantizeTableS));
+	MOVSS(XMM1, MDisp(EAX, (u32)(u64)m_quantizeTableS));
 	PUNPCKLDQ(XMM1, R(XMM1));
 	MULPS(XMM0, R(XMM1));
 	CVTPS2DQ(XMM0, R(XMM0));
@@ -309,7 +309,7 @@ void AsmRoutineManager::GenQuantizedStores() {
 
 	const u8* storePairedS16 = AlignCode4();
 	SHR(32, R(EAX), Imm8(6));
-	MOVSS(XMM1, MDisp(EAX, (u32)m_quantizeTableS));
+	MOVSS(XMM1, MDisp(EAX, (u32)(u64)m_quantizeTableS));
 	PUNPCKLDQ(XMM1, R(XMM1));
 	MULPS(XMM0, R(XMM1));
 	CVTPS2DQ(XMM0, R(XMM0));
